requirements.txt
Django==5.0.4
djangorestframework==3.15.1
djangorestframework_simplejwt==5.5.1
django-cors-headers==4.4.0
django-filter==25.1
django-taggit==5.0.1
django-seed==0.3.1
django-axes==8.0.0
dj-database-url==3.0.1
python-decouple==3.8
gunicorn==23.0.0
whitenoise==6.7.0
psycopg2-binary==2.9.10
Faker==37.5.3
python-dotenv
drf_spectacular
Pillow
